feat: 🎸 change to round flags in the Flag component
The previous rectangular flag svgs have been replaced with round svgs.
Alignment css should be the same as before. For users of this component,
the height props has been removed. Instead a prop called "size" is used.
It can take the values "s", "m" and "l" corresponding to a flag diameter
of 12px, 16px and 24px. This is a breaking change.

BREAKING CHANGE: 🧨 Replaced prop "height" with new prop "size".